Potential fix for code scanning alert no. 16: Implicit narrowing conversion in compound assignment

+++ src/main/java/cad/core/Sketch.java
@@ -1367,7 +1367,7 @@ public class Sketch {
                         float end = (float) Math.toRadians(arc.getEndAngle());
 
                         if (end < start) {
-                            end += 2.0 * Math.PI;
+                            end += (float) (2.0 * Math.PI);
                         }
 
                         float angleDiff = end - start;
